renamed connectionHandler and remoteHandler
