+++ handlers/settingsHandler.py
@@ -3,7 +3,7 @@ import msvcrt
 
 ## custom modules
 from handlers.local_handler import localHandler
-from handlers.remoteHandler import remoteHandler
+from handlers.remote_handler import remoteHandler
 
 from modules.score_rater import scoreRate
