+++ handlers/storageSettingsHandler.py
@@ -8,7 +8,7 @@ import time
 
 ## custom modules
 from handlers.local_handler import localHandler
-from handlers.remoteHandler import remoteHandler
+from handlers.remote_handler import remoteHandler
 
 class storageSettingsHandler():
