+++ seisen.py
@@ -9,7 +9,7 @@ import threading
 
 ## custom modules
 from handlers.local_handler import localHandler
-from handlers.remoteHandler import remoteHandler
+from handlers.remote_handler import remoteHandler
 
 from handlers.settingsHandler import settingsHandler
